windows/windowsservercore/hotspot/requirements.txt
pandas
gspread
oauth2client
google-analytics-data